build: allow resolving monorepo packages
This teaches jest, webpack, eslint, and typescript how to resolve monorepo packages. At the moment there are none except the "stub" `@votingworks/election-manager`, which is an app and is not intended to be imported anyway. The `tsconfig.json` at the root of the monorepo is the source of truth using the `compilerOptions.paths` option, and all the other tools use this to guide their own module resolution.

TypeScript just refers to `tsconfig.json` in each app/library, which must extend the root `tsconfig.json`. Eslint uses `eslint-import-resolver-typescript` to piggyback on TypeScript's resolver. Jest and webpack both get configured via `react-app-rewired` to allow us to customize the configuration in a non-ejected app built by `create-react-app`. Once we can edit their configurations, we use `tsconfig-paths` and `ts-jest` to teach jest how to resolve monorepo packages, and `tsconfig-paths-webpack-plugin` to teach webpack. Additionally, since babel does the actual TypeScript transpilation in a CRA app, we need to tell babel it's okay to transpile things outside of the application root of `apps/client/src`.